file structure changed and some variable names has been changed
