+++ src/App.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import Home from "./pages/Home";
-import LoginSigninHeader from "./pages/LoginSigninHeader";
-import ProtectedRoute from "./pages/protectedRoute";
-import Searchp from "./pages/Searchp";
+import Home from "./routes/Home";
+import LoginSigninHeader from "./routes/LoginSigninHeader";
+import ProtectedRoute from "./routes/ProtectedRoute";
+import Search from "./routes/Search";
 
 
 function App() {
@@ -12,8 +12,7 @@ function App() {
       <Switch>
         <ProtectedRoute path="/" component={Home} exact />
         <Route path="/login" component={LoginSigninHeader} exact />
-        
-        <ProtectedRoute path="/Searchp" component={Searchp} exact />
+        <ProtectedRoute path="/search" component={Search} exact />
         <Route path="*" render={() => (<div id="notfound">NOT FOUND 404</div>)} />
       </Switch>
     </BrowserRouter >
