+++ src/routes/Components/Header.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom'
 import Navbar from './Navbar';
 
 function Header() {
@@ -9,7 +10,7 @@ function Header() {
         <p>Having a car problem ?</p>
         <h1><span>We've Got You Covered</span></h1>
         <p className="details">If you're having any car related problems, and you want to fix it straight away, Chakika is the way to go, learn about Chakika, and how we serve our customers</p>
-        <a href="#" className="header-btn">About Us</a>
+        <Link to='#' className="header-btn">About Us</Link>
       </div>
     </div>
   );
