+++ src/routes/images/hero.png
