+++ src/styles/loginAndRegister.css
@@ -8,7 +8,7 @@
 .App__Aside__Logo {
   height: 49px;
   width: 158px;
-  background-image: url("../pages/images/logo1.png");
+  background-image: url("../routes/images/logo1.png");
   display: inline-block;
   margin-top: 10px;
   margin-left: 25px;
@@ -17,7 +17,7 @@
 .App__Aside {
   width: 50%;
   background-color: #fff;
-  background-image: url("../pages/images/header-bgfliped.png");
+  background-image: url("../routes/images/header-bgfliped.png");
 }
 
 .App__Form {
